Add files via upload
The aim of this model is identify corporate lending risk for banks or financial institution. The dataset contains information of 1384 Indian companies across 11 variables for the year 2019. Using this data, we are creating a set of supervised learning models with Logistic Regression, LDA and Random Forest to predict corporate default.